Extract operation semantics from 'vm_mngr' to dedicated 'op_semantics'
