+++ miasm2/jitter/arch/JitCore_aarch64.c
@@ -6,6 +6,7 @@
 #include "../vm_mngr.h"
 #include "../vm_mngr_py.h"
 #include "../JitCore.h"
+#include "../op_semantics.h"
 #include "JitCore_aarch64.h"
