+++ setup.py
@@ -32,26 +32,32 @@ def buil_all():
         Extension("miasm2.jitter.arch.JitCore_x86",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_x86.c"]),
         Extension("miasm2.jitter.arch.JitCore_arm",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_arm.c"]),
         Extension("miasm2.jitter.arch.JitCore_aarch64",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_aarch64.c"]),
         Extension("miasm2.jitter.arch.JitCore_msp430",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_msp430.c"]),
         Extension("miasm2.jitter.arch.JitCore_mips32",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_mips32.c"]),
         Extension("miasm2.jitter.arch.JitCore_ppc32",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_ppc32.c"],
                   depends=["miasm2/jitter/arch/JitCore_ppc32.h",
                            "miasm2/jitter/arch/JitCore_ppc32_regs.h"]),
@@ -68,26 +74,32 @@ def buil_all():
         Extension("miasm2.jitter.arch.JitCore_x86",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_x86.c"]),
         Extension("miasm2.jitter.arch.JitCore_arm",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_arm.c"]),
         Extension("miasm2.jitter.arch.JitCore_aarch64",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_aarch64.c"]),
         Extension("miasm2.jitter.arch.JitCore_msp430",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_msp430.c"]),
         Extension("miasm2.jitter.arch.JitCore_mips32",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_mips32.c"]),
         Extension("miasm2.jitter.arch.JitCore_ppc32",
                   ["miasm2/jitter/JitCore.c",
                    "miasm2/jitter/vm_mngr.c",
+                   "miasm2/jitter/op_semantics.c",
                    "miasm2/jitter/arch/JitCore_ppc32.c"],
                   depends=["miasm2/jitter/arch/JitCore_ppc32.h",
                            "miasm2/jitter/arch/JitCore_ppc32_regs.h"]),
